Rename the 'lib' directory to 'logging'
